Map testID to accessibility properties
On Android for the application elements to be targetable/found it needs
to use the `resourceId`, we assumed React Native would convert `testID`
to `resourceId`, unfortunately it didn't. With a quick search, you'll
find a few dozens of closed tickets on this subject.

This solution (current plugin) allow us developers to code as before,
completely oblivious of these `testID` differences between Android and
iOS.

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>