Support React Native Metro Preset
Since most of the users will be using the react-native preset for metro
this should be the main feature to be supported.

Due to all the plugins used by react-native preset and their order,
replacing the testID with the JSX spread was not working as expected
when transpiled to ES5 code.

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>